IMPORTANT: arnoldi iteration will now return a smaller decomposition than requested. arnoldi eigensolver is set up to handle this, so long as the subspace never shrinks below k dimensions. it seems wise, but it's a homebrew adjustment and thus probably bad, though tests suggest it works

+++ eigen_solvers.h
@@ -13,7 +13,7 @@ namespace eigen_solver {
 
   int qr_impshift_tridiag(const Eigen::MatrixXd& T, Eigen::MatrixXd& S, Eigen::VectorXd& thetas, const int maxiter, const double eigen_tol=1e-11, const double zero_tol=1e-13, const std::string shift_type="eigval")
 ;
-  int arnoldi_method_imprestart_hermitian(const Eigen::MatrixXd& A, const Eigen::VectorXd& v, Eigen::MatrixXd& V_ritz, Eigen::VectorXd& l_ritz, const int k, const int p, const double iram_maxiter, const int qr_maxiter, const double f_tol=1e-14, const double qr_eigen_tol=1e-11, const double qr_zero_tol=1e-13);
+  int arnoldi_method_imprestart_hermitian(const Eigen::MatrixXd& A, const Eigen::VectorXd& v, Eigen::MatrixXd& V_ritz, Eigen::VectorXd& l_ritz, const int k, const int p, const int iram_maxiter, const int qr_maxiter, const double f_tol=1e-14, const double qr_eigen_tol=1e-11, const double qr_zero_tol=1e-13);
 
 }
